test: restructure regressor tests
