--- tests/safeds/ml/regression/_ada_boost/test_fit.py
@@ -1,21 +1,0 @@
-import pytest
-from safeds.data.tabular.containers import Table, TaggedTable
-from safeds.exceptions import LearningError
-from safeds.ml.regression import AdaBoost
-from tests.fixtures import resolve_resource_path
-
-
-def test_ada_boost_fit() -> None:
-    table = Table.from_csv(resolve_resource_path("test_ada_boost.csv"))
-    tagged_table = TaggedTable(table, "T")
-    ada_boost = AdaBoost()
-    ada_boost.fit(tagged_table)
-    assert True  # This asserts that the fit method succeeds
-
-
-def test_ada_boost_fit_invalid() -> None:
-    table = Table.from_csv(resolve_resource_path("test_ada_boost_invalid.csv"))
-    tagged_table = TaggedTable(table, "T")
-    ada_boost = AdaBoost()
-    with pytest.raises(LearningError):
-        ada_boost.fit(tagged_table)
