--- tests/safeds/ml/regression/_ada_boost/test_predict.py
@@ -1,33 +1,0 @@
-import pytest
-from safeds.data.tabular.containers import Table, TaggedTable
-from safeds.exceptions import PredictionError
-from safeds.ml.regression import AdaBoost
-from tests.fixtures import resolve_resource_path
-
-
-def test_ada_boost_predict() -> None:
-    table = Table.from_csv(resolve_resource_path("test_ada_boost.csv"))
-    tagged_table = TaggedTable(table, "T")
-    ada_boost = AdaBoost()
-    ada_boost.fit(tagged_table)
-    ada_boost.predict(tagged_table.features)
-    assert True  # This asserts that the predict method succeeds
-
-
-def test_ada_boost_predict_not_fitted() -> None:
-    table = Table.from_csv(resolve_resource_path("test_ada_boost.csv"))
-    tagged_table = TaggedTable(table, "T")
-    ada_boost = AdaBoost()
-    with pytest.raises(PredictionError):
-        ada_boost.predict(tagged_table.features)
-
-
-def test_ada_boost_predict_invalid() -> None:
-    table = Table.from_csv(resolve_resource_path("test_ada_boost.csv"))
-    invalid_table = Table.from_csv(resolve_resource_path("test_ada_boost_invalid.csv"))
-    tagged_table = TaggedTable(table, "T")
-    invalid_tagged_table = TaggedTable(invalid_table, "T")
-    ada_boost = AdaBoost()
-    ada_boost.fit(tagged_table)
-    with pytest.raises(PredictionError):
-        ada_boost.predict(invalid_tagged_table.features)
